Prevent location decimals from mapping to wrong index.
Mapping indexes to embeddings works by multiplying y emb lenth by
y location and then adding x location.
EX:
point (14, 72)
Will map to the flattened embedding by 14 * 168 + 72=2424

The issue is when the location is a decimal
EX:
point (14.78, 72)
14.78 * 168 + 72 = 2555.04 was gets rounded to 2555
This is outside of y dim 14's "row" so 14 * 168 + 168 = 2520

This can also cause a runtime error if the final row has a point.
EX:
point (167.78, 72)
167.78 * 168 + 72 = 28259.04 which gets rounded down to 28259
This is out of bound for the flattened emb which is size 168 * 168 = 28224

+++ contraband/pipelines/contrastive_loss.py
@@ -56,12 +56,12 @@ def contrastive_volume_loss(emb_0, emb_1, locations_0, locations_1, t):
     ind_kernel = torch.Tensor([
         np.prod(volume_shape[i + 1:])
         for i in range(d)
-    ]).float().to(locations_0.device)
+    ]).long().to(locations_0.device)
 
     # turn point coordinates into indices
     # (n,) (TODO: concatenate batches here, do the same for embeddings)
-    ind_0 = torch.matmul(locations_0, ind_kernel).long().squeeze(dim=0)
-    ind_1 = torch.matmul(locations_1, ind_kernel).long().squeeze(dim=0)
+    ind_0 = torch.matmul(locations_0.long(), ind_kernel).squeeze(dim=0)
+    ind_1 = torch.matmul(locations_1.long(), ind_kernel).squeeze(dim=0)
 
     # get embeddings for each point
     # (b, n, c)
